Move Sentry initialization into config
