+++ src/main/kotlin/sh/zachwal/button/App.kt
@@ -24,6 +24,7 @@ import sh.zachwal.button.auth.configureFormAuth
 import sh.zachwal.button.auth.configureUserSessionAuth
 import sh.zachwal.button.auth.contact.ContactTokenCleanupTask
 import sh.zachwal.button.config.AppConfig
+import sh.zachwal.button.config.initSentry
 import sh.zachwal.button.controller.createControllers
 import sh.zachwal.button.features.configureRoleAuthorization
 import sh.zachwal.button.features.configureStatusPages
@@ -34,7 +35,6 @@ import sh.zachwal.button.guice.JdbiModule
 import sh.zachwal.button.guice.MessagingModule
 import sh.zachwal.button.roles.RoleAuthorization
 import sh.zachwal.button.roles.RoleService
-import sh.zachwal.button.sentry.initSentry
 import sh.zachwal.button.session.CONTACT_SESSION
 import sh.zachwal.button.session.CONTACT_SESSION_LENGTH
 import sh.zachwal.button.session.DbSessionStorage
